# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: "医疗智能问答系统 API"
  description: |
    # 医疗智能问答系统 RESTful API

    基于 Master-Agent 多智能体协作架构的医疗健康咨询系统。

    ## 核心特性
    - 多轮结构化问诊（InterviewAgent）
    - 异步循证知识检索（SearchAgent）
    - LLM综合推理诊断（DiagnosisAgent）
    - SSE流式事件推送
    - 红旗症状紧急中断

    ## 安全红线
    - 所有诊断输出含不确定性表达，**禁止AI确诊**
    - 红旗症状触发立即返回急救指引，**禁止继续问诊**
    - 用户PII信息在API入口层脱敏

    ## 免责声明
    本API返回的所有健康建议仅供参考，不能替代专业医疗诊断。

    ## 认证
    初期可选 `X-API-Key` 头部认证，生产环境必填。

  version: "1.0.0"
  contact:
    name: "开发团队"
    email: "dev@medical-qa.local"
servers:
  - url: "http://localhost:8000/api/v1"
    description: "本地开发环境"
  - url: "https://medical-qa-staging.example.com/api/v1"
    description: "测试环境"
  - url: "https://medical-qa.example.com/api/v1"
    description: "生产环境"
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API密钥（初期可选，生产环境必填）"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, code, message]
      properties:
        error:
          type: string
          description: "机器可读的错误码"
          example: "SESSION_NOT_FOUND"
        code:
          type: integer
          description: "HTTP状态码"
          example: 404
        message:
          type: string
          description: "人类可读的错误描述"
          example: "会话不存在或已被删除"
        details:
          type: object
          description: "附加错误详情"
        timestamp:
          type: string
          format: date-time
          example: "2026-06-19T10:30:00.000Z"
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: "总记录数"
          example: 150
        limit:
          type: integer
          description: "每页条数"
          example: 20
        offset:
          type: integer
          description: "偏移量"
          example: 0
        has_more:
          type: boolean
          description: "是否有更多数据"
          example: true

    # ============================================================
    # 会话 Schemas
    # ============================================================
    SessionStatus:
      type: string
      enum: [active, paused, completed, emergency_terminated, closed_timeout]
...
    WorkflowStage:
      type: string
      enum: [init, collect, diagnose, done, emergency]
...
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "m1b2c3d4-e5f6-7890-abcd-ef1234567890"
        session_id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        round_number:
          type: integer
          example: 1
        role:
          type: string
          enum: [user, assistant, system]
          example: "user"
        content:
          type: string
          description: "消息文本内容"
          example: "我头痛两天了，前额位置，大概7分痛"
        content_type:
          type: string
          enum: [text, question, diagnosis_report, emergency_guide, status_report]
          example: "text"
        agent_source:
          type: string
          enum: [master, interview, search, diagnosis, emergency, system]
          nullable: true
          example: null
        token_count:
          type: integer
          nullable: true
          example: null
        created_at:
          type: string
          format: date-time
          example: "2026-06-19T10:30:05.000Z"
    SendMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          description: "用户输入的原始消息（API入口层自动脱敏）"
          minLength: 1
          maxLength: 4096
          example: "我头痛两天了，前额位置，大概7分痛，还有点低烧"
        content_type:
          type: string
          enum: [text]
          default: "text"
    SendMessageResponse:
      type: object
      description: "发送消息的同步响应（含当前状态快照）"
      properties:
        message:
          $ref: "#/components/schemas/Message"
        session_status:
          $ref: "#/components/schemas/SessionStatus"
        current_stage:
          $ref: "#/components/schemas/WorkflowStage"
        red_flag_raised:
          type: boolean
          example: false
        round_count:
          type: integer
          example: 1
        collected_fields_summary:
          type: object
          description: "当前已采集的核心字段摘要"
          example:
            chief_complaint: "头痛"
            duration: "2天"
            location: "前额"
            severity: 7
        next_action:
          type: string
          enum: [continue, diagnosis_ready, emergency_interrupted, completed]
          description: |
            客户端后续行为指引：
            - continue: 继续问诊，连接SSE获取下一轮问题
            - diagnosis_ready: 信息充足，连接SSE获取诊断报告
            - emergency_interrupted: 红旗触发，会话已终止
            - completed: 问答/诊断已完成
          example: "continue"

    # ============================================================
    # SSE 事件 Schemas
    # ============================================================
...
  parameters:
    SessionIdParam:
      name: session_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: "会话ID"
    PaginationLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: "每页返回条数"
    PaginationOffset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: "偏移量（0-based）"

  # ============================================================
  # 通用响应
  # ============================================================
  responses:
    400BadRequest:
      description: "请求参数无效"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "VALIDATION_ERROR"
            code: 400
            message: "content字段为必填项且长度1-4096字符"
            timestamp: "2026-06-19T10:30:00.000Z"
...
    404NotFound:
      description: "资源不存在"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "SESSION_NOT_FOUND"
            code: 404
            message: "会话不存在或已被删除"
            timestamp: "2026-06-19T10:30:00.000Z"
    409Conflict:
      description: "状态冲突"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "SESSION_TERMINATED"
            code: 409
            message: "会话已紧急终止，无法继续操作"
            timestamp: "2026-06-19T10:30:00.000Z"
...
paths:
...
  /sessions/{session_id}/messages:
    post:
      tags: [Messages]
      summary: "发送用户消息（核心交互入口）"
      description: |
        向指定会话发送用户消息，触发Agent链处理。

        ## 处理流程
        1. 接收消息 → PII脱敏
        2. 红旗关键词检测
        3. 若触发红旗 → 返回 emergency_interrupted
        4. 若未触发 → MasterAgent路由 → 问诊/问答/紧急
        5. 返回同步响应 + next_action

        ## next_action 客户端行为指引
        | next_action | 含义 | 后续操作 |
        |------------|------|---------|
        | continue | 问诊进行中 | 连接SSE获取下一轮问题 |
        | diagnosis_ready | 信息充足 | 连接SSE获取诊断报告 |
        | emergency_interrupted | 红旗触发 | 会话已终止，读取emergency消息 |
        | completed | 问答/诊断完成 | 读取消息历史获取完整回答 |
      operationId: sendMessage
      parameters:
        - $ref: "#/components/parameters/SessionIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
            examples:
              initial_symptom:
                summary: "初次症状描述"
                value:
                  content: "我头痛两天了，前额位置，大概7分痛，还有点低烧"
                  content_type: "text"
              follow_up:
                summary: "追问回答"
                value:
                  content: "没有呕吐，对青霉素过敏"
                  content_type: "text"
              red_flag:
                summary: "红旗症状（触发紧急中断）"
                value:
                  content: "我胸口剧痛，呼吸困难，左手臂发麻"
                  content_type: "text"
      responses:
        "200":
          description: "消息接收成功，返回处理结果快照"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMessageResponse"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "404":
          $ref: "#/components/responses/404NotFound"
        "409":
          $ref: "#/components/responses/409Conflict"

    get:
      tags: [Messages]
      summary: "获取消息历史"
      description: "分页获取指定会话的对话消息历史，按时间正序排列"
      operationId: listMessages
      parameters:
        - $ref: "#/components/parameters/SessionIdParam"
        - name: round_number
          in: query
          schema:
            type: integer
          description: "按轮次筛选（可选）"
        - $ref: "#/components/parameters/PaginationLimit"
        - $ref: "#/components/parameters/PaginationOffset"
      responses:
        "200":
          description: "消息列表"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Message"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "404":
          $ref: "#/components/responses/404NotFound"

  # ==========================================================
  # SSE 流式事件推送
  # ==========================================================
  /sessions/{session_id}/stream:
    get:
      tags: [Streaming]
      summary: "SSE流式事件推送"
      description: |
        建立 Server-Sent Events 长连接，接收该会话的实时事件流。

        ## 事件类型路由表
        | event | 触发条件 | payload Schema | 客户端行为 |
        |-------|---------|----------------|-----------|
        | heartbeat | 每15秒无事件 | SSEHeartbeatEvent | 忽略，保活 |
        | message | 问诊Agent生成提问/直接回答 | SSEMessageEvent | 追加到对话界面 |
        | diagnosis_progress | 诊断Agent开始推理 | SSEDiagnosisProgressEvent | 更新进度条 |
        | diagnosis_complete | 诊断完成+Schema校验通过 | SSEDiagnosisCompleteEvent | 渲染报告，关闭连接 |
        | emergency | 任意阶段红旗触发 | SSEEmergencyEvent | **展示急救指引，锁定界面** |
        | error | 处理异常 | SSEErrorEvent | 展示错误，可重连 |

        ## 客户端示例 (JavaScript)
        ```javascript
        const es = new EventSource(`/api/v1/sessions/${sessionId}/stream`);
        es.addEventListener('message', (e) => { appendMessage(JSON.parse(e.data)); });
        es.addEventListener('diagnosis_progress', (e) => { updateProgress(JSON.parse(e.data)); });
        es.addEventListener('diagnosis_complete', (e) => { renderReport(JSON.parse(e.data).result); es.close(); });
        es.addEventListener('emergency', (e) => { showEmergency(JSON.parse(e.data)); es.close(); });
        ```
      operationId: streamEvents
      parameters:
        - $ref: "#/components/parameters/SessionIdParam"
      responses:
        "200":
          description: "SSE事件流"
          content:
            text/event-stream:
              schema:
                type: string
                description: "SSE事件流，每行格式: event: {type}\\ndata: {json}\\n\\n"
        "404":
          $ref: "#/components/responses/404NotFound"

  # ==========================================================
  # 诊断管理
  # ==========================================================
...